Modified structure a bit Added decision trees, will need to find better dataset
