+++ baselines/decision_trees/ObjectRecognition.py
@@ -1,0 +1,33 @@
+import numpy as np
+import pandas as pd
+from sklearn.datasets import load_iris
+from sklearn.model_selection import train_test_split
+from sklearn.tree import DecisionTreeClassifier
+import matplotlib.pyplot as plt
+from sklearn import tree
+
+# Load the Iris dataset
+iris = load_iris()
+X = iris.data
+y = iris.target
+
+# Split the dataset into training and testing sets
+X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
+
+# Create a decision tree classifier
+clf = DecisionTreeClassifier(random_state=42)
+
+# Fit the model on the training data
+clf.fit(X_train, y_train)
+
+# Make predictions on the test data
+y_pred = clf.predict(X_test)
+
+# Calculate accuracy
+accuracy = np.mean(y_pred == y_test)
+print(f'Accuracy: {accuracy:.2f}')
+
+# Visualize the decision tree (you can adjust the figsize as needed)
+plt.figure(figsize=(12, 8))
+tree.plot_tree(clf, feature_names=iris.feature_names, class_names=iris.target_names, filled=True)
+plt.show()
